refactor: move ResponseDTO, FailureResponseDTO to response folder
