+++ src/main/java/com/howaboutquestion/backend/domain/auth/service/AuthService.java
@@ -14,4 +14,6 @@ public class AuthService {
     public UserRegisterResponse basicRegister(UserRegisterRequest request){
         return userService.tryRegisterUser(request);
     }
+
+
 }
