+++ src/main/java/com/howaboutquestion/backend/global/error/GlobalExceptionHandler.java
@@ -2,20 +2,18 @@ package com.howaboutquestion.backend.global.error;
 
 
 import com.howaboutquestion.backend.global.common.StatusCode;
+import com.howaboutquestion.backend.global.response.FailureResponseDTO;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.NoHandlerFoundException;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.util.Objects;
 
 /**
  * packageName    : com.howaboutquestion.backend.global.error<br>
