+++ src/main/java/com/howaboutquestion/backend/global/response/FailureResponseDTO.java
@@ -1,7 +1,6 @@
-package com.howaboutquestion.backend.global.error;
+package com.howaboutquestion.backend.global.response;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.howaboutquestion.backend.global.common.ResponseDTO;
 import com.howaboutquestion.backend.global.common.StatusCode;
 import lombok.*;
