+++ src/main/java/com/howaboutquestion/backend/global/util/ResponseUtility.java
@@ -1,7 +1,7 @@
 package com.howaboutquestion.backend.global.util;
 
 import com.howaboutquestion.backend.global.common.StatusCode;
-import com.howaboutquestion.backend.global.error.FailureResponseDTO;
+import com.howaboutquestion.backend.global.response.FailureResponseDTO;
 import com.howaboutquestion.backend.global.response.SuccessResponseDTO;
 import org.springframework.http.ResponseEntity;
